% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_install_package.R
\name{build_and_install_package}
\alias{build_and_install_package}
\title{Build, Install and Load Package}
\usage{
build_and_install_package(folder_source_packages, pkgname_pkgversion,
  verbose = FALSE)
}
\arguments{
\item{folder_source_packages}{\code{character} directory containing local
source packages, that will be built into the Docker image.}

\item{pkgname_pkgversion}{\code{character} the package name concatenated
with the version number of the package.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
the package is built, install and loaded as side effects.
Furthermore the source package is saved in the folder containing the files
for the Docker image.
}
\description{
Builds, installs and load the package. As a side effect the source package
is built and saved into the directory with the Docker files.
}
