% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_deps.R
\name{combine_deps}
\alias{combine_deps}
\title{Combine Identified Dependencies with CRAN and Local Source Packages}
\usage{
combine_deps(pkg_deps, deps_cran, deps_local, prioritize_cran = FALSE)
}
\arguments{
\item{pkg_deps}{\code{data.frame} all package dependencies with package name
and version number.}

\item{deps_cran}{\code{data.frame} package dependencies matched with
CRAN.}

\item{deps_local}{\code{data.frame} package dependencies matched with
local source packages.}

\item{prioritize_cran}{\code{logical} should R dependency packages matched
with CRAN be prioritized over matches with local source packages?}
}
\value{
\code{list} prioritized dependencies split into CRAN dependencies
and dependencies of local source packages.
}
\description{
Combine Identified Dependencies with CRAN and Local Source Packages
}
