% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_compose.r
\name{compose_up}
\alias{compose_up}
\title{Pull, create and start all conainters defined in a Docker Compose file}
\usage{
compose_up(compose, force_install = FALSE)
}
\arguments{
\item{compose}{Path to a Docker Compose file. You can also parse the file
first, change it and supply it as an object.}

\item{force_install}{If TRUE, removes all containers and re-creates them from
the compose file. If 2, the images are also re-downloaded. Danger: this
will destroy the indexes in your containers!}
}
\description{
Pull, create and start all conainters defined in a Docker Compose file
}
