% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_local_pkgs.R
\name{copy_local_pkgs}
\alias{copy_local_pkgs}
\title{Copy Local Source Packages to Docker Subdirectory}
\usage{
copy_local_pkgs(pkgs_df = NULL, dir_src = NULL,
  dir_src_docker = NULL, verbose = TRUE)
}
\arguments{
\item{pkgs_df}{\code{data.frame} with names and version numbers of packages
to be copied to Docker subdirectory.}

\item{dir_src}{\code{character} directories with local source packages. Put
directories in prioritized order. The first directory will have the highest}

\item{dir_src_docker}{\code{character} directory of subfolder with
relevant source packages for a Docker image. Source packages will be copied
to this directory.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
invisibly. As a side effect source package files (*.tar.gz) are
copied to Docker subdirectory.
}
\description{
Copy Local Source Packages to Docker Subdirectory
}
