% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_from_statement.R
\name{create_from_statement}
\alias{create_from_statement}
\title{Create From Statement for Dockerfile}
\usage{
create_from_statement(r_version = NULL, verbose = FALSE)
}
\arguments{
\item{r_version}{\code{character} which version of base R to include in the 
Docker image, e.g. '3.6.0'. Defaults to NULL, which implies that the active 
version of R will apply.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{character} FROM statement for Dockerfile.
}
\description{
Creates FROM statement for Dockerfile and sets it up with a relevant
`rocker` r-base image as a starting point.
}
