% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statement_cran_versions.R
\name{create_statement_cran_versions}
\alias{create_statement_cran_versions}
\title{Create Install CRAN Packages Statement for Dockerfile}
\usage{
create_statement_cran_versions(deps, verbose = TRUE)
}
\arguments{
\item{deps}{\code{data.frame} dependency packages with version numbers.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{character} statement(s) for Dockerfile.
}
\description{
Create Install CRAN Packages Statement for Dockerfile
}
