% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statement_local_pkgs.R
\name{create_statement_local_pkgs}
\alias{create_statement_local_pkgs}
\title{Create Install Statements for Local Source Packages}
\usage{
create_statement_local_pkgs(pkgs_df, dir_src_docker = "",
  verbose = TRUE)
}
\arguments{
\item{pkgs_df}{\code{data.frame} with names and version numbers of desired
packages.}

\item{dir_src_docker}{\code{character} subdirectory of Docker image
directory with relevant source packages, that will be installed in the Docker
container. Relevant local source packages will be copied to this directory.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{character} install chunk for Dockerfile.
}
\description{
Create Install Statements for Local Source Packages
}
