% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statement_main_package.R
\name{create_statement_main_package}
\alias{create_statement_main_package}
\title{Create Install Statement for Main Package}
\usage{
create_statement_main_package(dir_source_packages, pkgname_pkgversion,
  verbose = TRUE)
}
\arguments{
\item{dir_source_packages}{\code{character} directory of subfolder with
relevant source packages for a Docker image.}

\item{pkgname_pkgversion}{\code{character} package name and package version
number for main package - concatenated.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{character} chunk with install statement for Dockerfile.
}
\description{
Create Install Statement for Main Package
}
