% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.r
\name{docker_exec}
\alias{docker_exec}
\title{Execute commands inside a running container}
\usage{
docker_exec(id, cmd)
}
\arguments{
\item{id}{ID or name of container.}

\item{cmd}{command to be run.}
}
\description{
Execute commands inside a running container
}
\examples{
\dontrun{
docker_exec(id = "amcat4",
            cmd = "amcat4 create-admin --username admin --password supergeheim")

docker_exec(id = "amcat4",
            cmd = "amcat4 create-test-index")
}
}
