% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{docker_inspect_container}
\alias{docker_inspect_container}
\title{Inspect Docker container}
\usage{
docker_inspect_container(id, parse = FALSE)
}
\arguments{
\item{id}{ID or name of container.}

\item{parse}{parse json string}
}
\description{
Retrieves existing container as json string
}
