% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{docker_lc}
\alias{docker_lc}
\title{List docker containers}
\usage{
docker_lc(..., all = FALSE)
}
\arguments{
\item{...}{named values are used as filters.}

\item{all}{include stopped containers.}
}
\description{
List docker containers
}
\examples{
\dontrun{
docker_lc(name = c("amcat", "elastic"))
}
}
