% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{docker_li}
\alias{docker_li}
\title{List docker images}
\usage{
docker_li(...)
}
\arguments{
\item{...}{named values are used as filters.}
}
\description{
List docker images
}
\examples{
\dontrun{
docker_li(reference = "amcat4docker")
}
}
