% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{docker_ping}
\alias{docker_ping}
\title{Ping Docker daemon}
\usage{
docker_ping()
}
\description{
Fails if Docker daemon is not reachable. On Unix systems, you want to either
run Docker in rootless mode (
\href{https://docs.docker.com/engine/install/linux-postinstall/}{see the
Docker documentation}) or start R with sudo for the correct priviliges. On
Windows, you should expose daemon on tcp://localhost:2375 without TLS (e.g.,
from Docker Desktop setting).
}
