% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.r
\name{docker_rmc}
\alias{docker_rmc}
\title{Remove a Docker container}
\usage{
docker_rmc(container, force = FALSE, link = FALSE)
}
\arguments{
\item{container}{ID or name of the container.}

\item{force}{If the container is running, kill it before removing it.}

\item{link}{Remove the specified link associated with the container.}
}
\description{
Remove a Docker container
}
