% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.r
\name{docker_rmi}
\alias{docker_rmi}
\title{Remove a Docker image}
\usage{
docker_rmi(image, force = FALSE, noprune = FALSE)
}
\arguments{
\item{image}{Image name or ID}

\item{force}{Remove the image even if it is being used by stopped containers or has other tags}

\item{noprune}{Do not delete untagged parent images}
}
\description{
Remove a Docker image
}
