% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.r
\name{docker_stop}
\alias{docker_stop}
\title{Stop container}
\usage{
docker_stop(id, kill_after = 60L)
}
\arguments{
\item{id}{ID or name of the container}

\item{kill_after}{Number of seconds to wait before killing the container}
}
\description{
Stop container
}
