% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockr.R
\name{dockr}
\alias{dockr}
\title{dockr class}
\description{
A R6 Docker class as a wrapper for the docker commands.

Docker System Wrapper
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dockr-new}{\code{dockr$new()}}
\item \href{#method-dockr-print}{\code{dockr$print()}}
\item \href{#method-dockr-show_output}{\code{dockr$show_output()}}
\item \href{#method-dockr-show_json}{\code{dockr$show_json()}}
\item \href{#method-dockr-clone}{\code{dockr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dockr-new"></a>}}
\if{latex}{\out{\hypertarget{method-dockr-new}{}}}
\subsection{Method \code{new()}}{
A Docker wrapper to run system commands for image management
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dockr$new(process = NA, commands = NULL, options = NULL, args = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process}}{The system process. Plan is to add \code{Podman} support but currently only
setup for \code{docker}. Although Podman commands are formatted in a similar context.}

\item{\code{commands}}{The main docker commands. Currently support commands
are "image", "container", "search", "build", "history", "inspect".}

\item{\code{options}}{the \code{command} option arguments or \code{image}}

\item{\code{args}}{Additional arguments or flags for the docker commands}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dockr-print"></a>}}
\if{latex}{\out{\hypertarget{method-dockr-print}{}}}
\subsection{Method \code{print()}}{
Prints the full command to the console for viewing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dockr$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dockr-show_output"></a>}}
\if{latex}{\out{\hypertarget{method-dockr-show_output}{}}}
\subsection{Method \code{show_output()}}{
Displays returned output in \code{tibble} format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dockr$show_output()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dockr-show_json"></a>}}
\if{latex}{\out{\hypertarget{method-dockr-show_json}{}}}
\subsection{Method \code{show_json()}}{
Show the json in pretty format from the command
Formatting Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dockr$show_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dockr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dockr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dockr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
