% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actual_package_versions.R
\name{get_actual_package_versions}
\alias{get_actual_package_versions}
\title{Get Version Numbers of Loaded and Installed Packages}
\usage{
get_actual_package_versions(pkgs)
}
\arguments{
\item{pkgs}{\code{character} names of packages.}
}
\value{
\code{data.frame} with packages and relevant version numbers.
}
\description{
Gets version numbers of specific packages. If a package is loaded, the version
number of the loaded package is used. If a packages is not loaded but
installed, the version number of the installed package will be extracted
in sted.
}
