% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loaded_packages_versions.R
\name{get_loaded_packages_versions}
\alias{get_loaded_packages_versions}
\title{Get Version Numbers of Loaded Packages}
\usage{
get_loaded_packages_versions(pkgs)
}
\description{
Get Version Numbers of Loaded Packages
}
