% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_dependencies.R
\name{identify_dependencies}
\alias{identify_dependencies}
\title{Identify Package Dependencies}
\usage{
identify_dependencies(dependencies = c("Depends", "Imports",
  "LinkingTo"), installed = TRUE, available = TRUE, base = FALSE,
  recommended = FALSE)
}
\arguments{
\item{dependencies}{character vector of dependency types to include.
    Choices are "Depends", "Imports", "LinkingTo", "Suggests",
    and "Enhances".  Defaults to \code{c("Depends", "Imports",
      "LinkingTo")}.
  }

\item{installed}{Logical indicating whether to pull dependency
    information from installed packages. Defaults to TRUE.
  }

\item{available}{Logical indicating whether to pull dependency
    information from available packages.  Defaults to TRUE.
  }

\item{base}{Logical indicating whether to include dependencies on base
    packages that are included in the R installation. Defaults to
    FALSE.
    }

\item{recommended}{Logical indicating whether to include dependencies on
    recommended packages that are included in the R
    installation. Defaults to
    FALSE.
    }
}
\value{
\code{character} all dependencies.
}
\description{
Identify Package Dependencies
}
