% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pkg_cran.R
\name{match_pkg_archive}
\alias{match_pkg_archive}
\title{Match Archived Versions of CRAN Packages}
\usage{
match_pkg_archive(pkgs_df)
}
\arguments{
\item{pkgs_df}{\code{data.frame} with names of R packages "pkg" and their
corresponding version numbers "vrs".}
}
\value{
\code{data.frame} with package names (and versions) of dependency
packages matched with CRAN repos.
}
\description{
Searches for a specific _archived_ version of an R package dependency
on CRAN.
}
