% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pkg_cran.R
\name{match_pkg_cran}
\alias{match_pkg_cran}
\title{Match Packages with CRAN}
\usage{
match_pkg_cran(pkgs_df, verbose = TRUE)
}
\arguments{
\item{pkgs_df}{\code{data.frame} with names of R packages "pkg" and their
corresponding version numbers "vrs".}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{data.frame} with package names (and versions) of dependency
packages matched with CRAN repos.
}
\description{
Match specific package versions with CRAN.
}
