% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pkg_local.R
\name{match_pkg_local}
\alias{match_pkg_local}
\title{Match Package with Local Source Packages}
\usage{
match_pkg_local(pkgs_df = NULL, dir_src = NULL,
  dir_src_docker = NULL, verbose = TRUE)
}
\arguments{
\item{pkgs_df}{\code{data.frame} with packages and their respective version
numbers to be matched.}

\item{dir_src}{\code{character} directories with local source packages. 
Note, the source packages must have filenames like
(packageName)_(packageVersion).tar.gz, e.g. "recorder_0.8.2.tar.gz".
Put directories in prioritized order. The first directory will have the 
highest priority.}

\item{dir_src_docker}{\code{character} subdirectory of Docker image
directory with relevant source packages, that will be installed in the Docker
container. Relevant local source packages will be copied to this directory.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{data.frame} with packages, their version numbers and the
directory, where the source package was found.
}
\description{
Match Package with Local Source Packages
}
