% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_package_versions.R
\name{mirror_package_versions}
\alias{mirror_package_versions}
\title{Mirror Version Numbers of Loaded and Installed Packages}
\usage{
mirror_package_versions(pkgs)
}
\arguments{
\item{pkgs}{\code{character} names of packages.}
}
\value{
\code{data.frame} with package names and their relevant version 
numbers.
}
\description{
Mirrors version numbers of specific packages. If a package is loaded, the
version number of the loaded package is used. If a packages is not loaded but
installed, the version number of the installed package will be extracted
in stead.
}
