% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_yml}
\alias{parse_yml}
\title{Parse Docker Compose Yaml file}
\usage{
parse_yml(x)
}
\arguments{
\item{x}{A yaml/yml Docker Compose file}
}
\value{
A data.frame with image/container configurations
}
\description{
Parse Docker Compose Yaml file
}
\examples{
parse_yml(system.file("extdata/docker-compose.yml", package = "dockr"))
}
