% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_docker.R
\name{prep_docker}
\alias{prep_docker}
\title{Prepare Docker folder}
\usage{
prep_docker(directory = NULL)
}
\arguments{
\item{directory}{\code{character} path for Docker files.}
}
\value{
\code{list} directories and paths for Docker files.
}
\description{
Prepares Docker folder in package directory.
}
