% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_docker_image.R
\name{prepare_docker_image}
\alias{prepare_docker_image}
\title{Prepare Docker Image}
\usage{
prepare_docker_image(directory = NULL, print_dockerfile = FALSE,
  verbose = TRUE, r_version = NULL, dependencies = c("Depends",
  "Imports", "LinkingTo"), base = FALSE, recommended = FALSE,
  dir_src = NULL)
}
\arguments{
\item{directory}{\code{character} directory where the files for the Docker
image are put.}

\item{print_dockerfile}{\code{logical} should the resulting Dockerfile be
printed?}

\item{verbose}{\code{logical} should messages be printed or not?}

\item{r_version}{\code{character} which version of R should the Docker image
run in, e.g. '3.6.0'. If set to NULL, the active version will apply.
#' @param dir_src \code{character} directories with local source packages. Put
directories in prioritized order. The first directory will have the highest
priority.}

\item{dependencies}{character vector of dependency types to include.
    Choices are "Depends", "Imports", "LinkingTo", "Suggests",
    and "Enhances".  Defaults to \code{c("Depends", "Imports",
      "LinkingTo")}.
  }

\item{base}{Logical indicating whether to include dependencies on base
    packages that are included in the R installation. Defaults to
    FALSE.
    }

\item{recommended}{Logical indicating whether to include dependencies on
    recommended packages that are included in the R
    installation. Defaults to
    FALSE.
    }
}
\value{
\code{list} relevant meta data of the files, that constitute
the Docker image. As a side effect the files for the Docker image -
including the resulting Dockerfile - are saved in the desired directory.
}
\description{
Prepares Docker image for the package by (1) looking up its dependencies,
(2) linking them to the right repositories and finally (3) creating all of
the necessary files for the image - including the Dockerfile.
}
