% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_end_service_information.R
\name{print_end_service_information}
\alias{print_end_service_information}
\title{Print Service Information}
\usage{
print_end_service_information(path_Dockerfile, dir_image, pkgname_pkgvrs)
}
\arguments{
\item{path_Dockerfile}{\code{character} path for Dockerfile.}

\item{dir_image}{\code{character} path for files for Docker image.}

\item{pkgname_pkgvrs}{\code{character} package name concatenated with
package version number.}
}
\value{
returns invisibly.
}
\description{
Prints service information, when files for Docker image have been created.
}
