% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_file.R
\name{print_file}
\alias{print_file}
\title{Print Contents of Text File}
\usage{
print_file(path)
}
\arguments{
\item{path}{\code{character} file path of text file.}
}
\description{
Prints all contents of text file.
}
