% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lib_paths.R
\name{set_lib_paths}
\alias{set_lib_paths}
\title{Set Library Paths Temporarily}
\usage{
set_lib_paths(dir_install)
}
\arguments{
\item{dir_install}{\code{character} where should the package be installed
on your system. Choose from `auto` (automatic detection), `temp` 
(temporary directory) or specify directory yourself.}
}
\value{
invisibly. Sets the library paths temporarily as a side effect. 
This will determine, where the package is installed.
}
\description{
Set Library Paths Temporarily
}
