% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rocker_image.R
\name{set_rocker_image}
\alias{set_rocker_image}
\title{Set Rocker Image}
\usage{
set_rocker_image(r_version = NULL)
}
\arguments{
\item{r_version}{\code{character} desired version of R. If NULL, the active
version of R will be used.}
}
\value{
\code{character} string with FROM statement for Dockerfile.
}
\description{
Set rocker image to reflect desired version of R.
}
