% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir_image.R
\name{setup_dir_image}
\alias{setup_dir_image}
\title{Setup Directory for Docker Image}
\usage{
setup_dir_image(pkg = ".", dir_image = NULL, overwrite = TRUE,
  verbose = FALSE)
}
\arguments{
\item{pkg}{Path to a package, or within a package.}

\item{dir_image}{\code{character} directory where to create folder for 
all files for the Docker image, from where the Docker image can be build.}

\item{overwrite}{\code{logical} if the directory for the files for the
Docker image already exists and is non-empty, should it be 
deleted/overwritten?}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{list} directories and paths for files for Docker image.
}
\description{
Sets up the desired directory with the appropriate file structure for the
Docker image and an empty Dockerfile.
}
