% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_directory_for_docker.R
\name{setup_directory_for_docker}
\alias{setup_directory_for_docker}
\title{Setup Directory for Docker Image}
\usage{
setup_directory_for_docker(directory = NULL, verbose = FALSE)
}
\arguments{
\item{directory}{\code{character} directory where the files for the Docker
image are put.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{list} directories and paths for Docker files.
}
\description{
Prepares the desired directory with the appropriate file structure for the
Docker image.
}
