% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_from_statement.R
\name{create_from_statement}
\alias{create_from_statement}
\title{Create From Statement for Dockerfile}
\usage{
create_from_statement(r_version = NULL, verbose = FALSE)
}
\arguments{
\item{r_version}{\code{character} which version of R should the Docker image
run in, e.g. '3.6.0'. If set to NULL, the active version will apply.
#' @param dir_src \code{character} directories with local source packages. Put
directories in prioritized order. The first directory will have the highest
priority.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{character} FROM statement for Dockerfile.
}
\description{
Creates FROM statement for Dockerfile and sets it up with a relevant
`rocker` image as a starting point.
}
