% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pkg_local.R
\name{match_pkg_local}
\alias{match_pkg_local}
\title{Match Package with Local Source Packages}
\usage{
match_pkg_local(pkgs_df = NULL, dir_src = NULL,
  dir_src_docker = NULL, verbose = TRUE)
}
\arguments{
\item{pkgs_df}{\code{data.frame} with packages and their respective version
numbers to be matched.}

\item{dir_src_docker}{\code{character} directory of subfolder with
relevant source packages for a Docker image. Source packages will be copied
to this directory.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{data.frame} with packages, their version numbers and the
directory, where the source package was found.
}
\description{
Match Package with Local Source Packages
}
