% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir_image.R
\name{setup_dir_image}
\alias{setup_dir_image}
\title{Setup Directory for Docker Image}
\usage{
setup_dir_image(pkg = ".", directory = NULL, verbose = FALSE)
}
\arguments{
\item{pkg}{Path to a package, or within a package.}

\item{directory}{\code{character} directory where the files for the Docker
image are put, and from where the Docker image can be build.}

\item{verbose}{\code{logical} should messages be printed or not?}
}
\value{
\code{list} directories and paths for Docker files.
}
\description{
Prepares the desired directory with the appropriate file structure for the
Docker image and an empty Dockerfile.
}
