% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines_to_file.R
\name{write_lines_to_file}
\alias{write_lines_to_file}
\title{Write Lines to File}
\usage{
write_lines_to_file(lines,
  filepath = "/home/w19799@CCTA.DK/projects/dockr_0.8.0/Dockerfile",
  prepend = FALSE, print_file = TRUE)
}
\arguments{
\item{lines}{\code{character} lines that will be written to file.}

\item{filepath}{\code{character} filepath of file to write lines to.}

\item{prepend}{\code{logical} should the new lines be prepended to
the existing text? Defaults to FALSE, which implies that the new lines will
be appended.}

\item{print_file}{\code{logical} should the contents of the file be printed
after editing? Defaults to TRUE.}
}
\value{
invisible return.
}
\description{
Write Lines to File
}
